+++ tests/test-listing.php
@@ -52,16 +52,32 @@ class AZ_Listing_Tests extends WP_UnitTestCase {
 	}
 
 	function test_populated_listing() {
+		$title = 'Test Page';
 		$p = $this->factory->post->create( array(
-			'post_title' => 'Test Page',
+			'post_title' => $title,
 			'post_type' => 'page',
 		) );
 		$q = new WP_Query( array(
 			'post_type' => 'page',
 		) );
 
-		$expected = sprintf( file_get_contents( 'tests/populated-listing.txt' ), $p );
+		$expected = sprintf( file_get_contents( 'tests/populated-listing.txt' ), $title, $p );
 		$actual = get_the_a_z_listing( $q );
+
+		$expected = preg_replace( '/\s{2,}|\t|\n/', '', $expected );
+		$actual = preg_replace( '/\s{2,}|\t|\n/', '', $actual );
+		$this->assertEquals( $expected, $actual );
+	}
+
+	function test_populated_taxonomy_listing() {
+		$title = 'Test Category';
+		$t = $this->factory->term->create( array(
+			'name' => $title,
+			'taxonomy' => 'category',
+		) );
+
+		$expected = sprintf( file_get_contents( 'tests/populated-taxonomy-listing.txt' ), $title, $t );
+		$actual = get_the_a_z_listing( 'category' );
 
 		$expected = preg_replace( '/\s{2,}|\t|\n/', '', $expected );
 		$actual = preg_replace( '/\s{2,}|\t|\n/', '', $actual );
@@ -103,19 +119,36 @@ class AZ_Listing_Tests extends WP_UnitTestCase {
 	}
 
 	function test_populated_lowercase_listing() {
+		$title = 'test page';
 		$p = $this->factory->post->create( array(
-			'post_title' => 'test page',
+			'post_title' => $title,
 			'post_type' => 'page',
 		) );
 		$q = new WP_Query( array(
 			'post_type' => 'page',
 		) );
 
-		$expected = sprintf( file_get_contents( 'tests/populated-listing-lowercase.txt' ), $p );
+		$expected = sprintf( file_get_contents( 'tests/populated-listing.txt' ), $title, $p );
 		$actual = get_the_a_z_listing( $q );
 
 		$expected = preg_replace( '/\s{2,}|\t|\n/', '', $expected );
 		$actual = preg_replace( '/\s{2,}|\t|\n/', '', $actual );
+		$this->assertEquals( $expected, $actual );
+	}
+
+	function test_populated_lowercase_taxonomy_listing() {
+		$title = 'test category';
+		$t = $this->factory->term->create( array(
+			'name' => $title,
+			'taxonomy' => 'category',
+		) );
+
+		$expected = sprintf( file_get_contents( 'tests/populated-taxonomy-listing.txt' ), $title, $t );
+		$actual = get_the_a_z_listing( 'category', false );
+
+		$expected = preg_replace( '/\s{2,}|\t|\n/', '', $expected );
+		$actual = preg_replace( '/\s{2,}|\t|\n/', '', $actual );
+
 		$this->assertEquals( $expected, $actual );
 	}
 }