+++ changelog.md
@@ -1,3 +1,15 @@
+## Full Changelog
+
+### 1.7.0
+* Add support for taxonomy term listings to the shortcode
+* Add support for filtering by taxonomy terms to the shortcode
+
+### 1.6.5
+* Regression fix for widget accessing WP_Post object as array
+
+### 1.6.4
+* Bugfix for accessing array as object PHP Warning. Reported by @babraham76
+
 ### 1.6.2
 * Bugfix for more complex templates - accessing post thumbnails failed.